requirements.txt
torch-scatter==2.0.8
torchmetrics==0.4.1
anytree==2.8.0
rdflib==6.1.1
